feat: Add method to retrieve invoices for a client
This commit adds a new method, `consultarFacturas`, to the `Cliente` class. This method allows retrieving a list of invoices associated with a client from the database. The method executes a SQL query to fetch the invoice information and returns a list of strings containing the invoice details. If no invoices are found for the client, a default message is added to the list.

The new method enhances the functionality of the `Cliente` class by providing a convenient way to retrieve invoice information for a client.

Refactor the code to add the `consultarFacturas` method to the `Cliente` class.